requirements.txt
opencv-python
numpy
python-telegram-bot